(nano) v1.3.0 - Add db.emit() function
db.emit can be used to emit a document to a supercouch view from the client.
This is not recommended except as an optimisation, for example to enrich an
existing supercouch entry.

Here's an example use case. We emit a value that describes the state of an
entity. Part of that is a cost, that can sometimes be known, sometimes not.
It's possible to compute the cost using complex requests for adjacent data.
However doing this everytime makes no sense, it's way better to enrich the
transaction state after the cost has been computed, so we re-emit the same
document to the view, with `score = score + 1`. This way, the view now includes
the entity with its cost, making subsequent requests faster.

Upgrade to sset.redis:0.8.1